+++ packages/ThemeDefault/src/partial/constant.latte
@@ -1,5 +1,5 @@
 <td class="attributes">
-    <code>{$constant->typeHint|typeLinks:$constant|noescape}</code>
+    <code>{$constant->getTypeHint()|typeLinks:$constant|noescape}</code>
 </td>
 <td class="name">
     <div class="description detailed">
@@ -21,6 +21,6 @@
 <td class="value">
     <div>
         <a href="#{$constant->getName()}" class="anchor">#</a>
-        <code>{$constant->valueDefinition|highlightValue:$class|noescape}</code>
+        <code>{$constant->getValueDefinition()|highlightValue:$class|noescape}</code>
     </div>
 </td>
